+++ app/javascript/apps/ArtsyPullRequests/index.ts
@@ -1,0 +1,1 @@
+export * from './ArtsyPullRequests'
