+++ app/javascript/apps/DumReader/components/NormalModeHelp/index.tsx
