+++ app/javascript/apps/DumReader/components/VisualModeHelp/index.tsx
