+++ app/javascript/apps/DumReader/index.ts
@@ -1,0 +1,1 @@
+export * from './DumReader'
