+++ app/javascript/apps/DumReader/shared/Entry.spec.ts
