+++ app/javascript/apps/DumReader/shared/Entry.ts
