+++ app/javascript/apps/DumReader/shared/Reader.ts
