+++ app/javascript/apps/DumReader/shared/colors.ts
