+++ app/javascript/apps/Projects/index.ts
@@ -1,0 +1,1 @@
+export * from './Projects'
