+++ app/javascript/apps/Projects/shared/colors.ts
