+++ app/views/projects/index.html.haml
@@ -7,6 +7,6 @@
   :javascript
     window.projects = #{projects.to_json.html_safe}
 
-  %main#root= javascript_pack_tag 'application'
+  %main#root= javascript_pack_tag 'projects'
 
   %h3 Try hard and believe in yourself!
